requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.2
MarkupSafe==3.0.2
multidict==6.3.2
openai==0.28.0
oscrypto==1.3.0
pillow==11.1.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
yarl==1.19.0
